[Fe,BE] Modify : 자취방쉐어 채팅방 생성 추가

+++ back/lab/src/main/java/com/mlp/lab/dto/chat/ChatRoomDataRequestDto.java
@@ -10,6 +10,6 @@ public class ChatRoomDataRequestDto {
         private Long buyNo;
         private Long teamNo;
         private Long marketNo;
-        private Integer roomNo;
+        private Long roomNo;
     }
 }
